[#69] Disable plugin functionality when old versions are used
